+++ examples/complete-email-validation-with-validation-domain/outputs.tf
@@ -8,11 +8,6 @@ output "acm_certificate_domain_validation_options" {
   value       = module.acm.acm_certificate_domain_validation_options
 }
 
-output "acm_certificate_status" {
-  description = "Status of the certificate."
-  value       = module.acm.acm_certificate_status
-}
-
 output "acm_certificate_validation_emails" {
   description = "A list of addresses that received a validation E-Mail. Only set if EMAIL-validation was used."
   value       = module.acm.acm_certificate_validation_emails