+++ main.tf
@@ -1,23 +1,23 @@
 locals {
-  create_certificate = var.create_certificate && var.putin_khuylo
+  create_certificate          = var.create_certificate && var.putin_khuylo
+  create_route53_records_only = var.create_route53_records_only && var.putin_khuylo
 
   # Get distinct list of domains and SANs
-  distinct_domain_names = distinct(
+  distinct_domain_names = coalescelist(var.distinct_domain_names, distinct(
     [for s in concat([var.domain_name], var.subject_alternative_names) : replace(s, "*.", "")]
-  )
+  ))
 
   # Get the list of distinct domain_validation_options, with wildcard
   # domain names replaced by the domain name
-  validation_domains = local.create_certificate ? distinct(
-    [for k, v in aws_acm_certificate.this[0].domain_validation_options : merge(
+  validation_domains = local.create_certificate || local.create_route53_records_only ? distinct(
+    [for k, v in try(aws_acm_certificate.this[0].domain_validation_options, var.acm_certificate_domain_validation_options) : merge(
       tomap(v), { domain_name = replace(v.domain_name, "*.", "") }
     )]
   ) : []
 }
 
 resource "aws_acm_certificate" "this" {
-  provider = aws.acm
-  count    = local.create_certificate ? 1 : 0
+  count = local.create_certificate ? 1 : 0
 
   domain_name               = var.domain_name
   subject_alternative_names = var.subject_alternative_names
@@ -25,6 +25,15 @@ resource "aws_acm_certificate" "this" {
 
   options {
     certificate_transparency_logging_preference = var.certificate_transparency_logging_preference ? "ENABLED" : "DISABLED"
+  }
+
+  dynamic "validation_option" {
+    for_each = var.validation_option
+
+    content {
+      domain_name       = try(validation_option.value["domain_name"], validation_option.key)
+      validation_domain = validation_option.value["validation_domain"]
+    }
   }
 
   tags = var.tags
@@ -35,8 +44,7 @@ resource "aws_acm_certificate" "this" {
 }
 
 resource "aws_route53_record" "validation" {
-  provider = aws.dns
-  count    = local.create_certificate && var.validation_method == "DNS" && var.create_route53_records && var.validate_certificate ? length(local.distinct_domain_names) : 0
+  count = (local.create_certificate || local.create_route53_records_only) && var.validation_method == "DNS" && var.create_route53_records && (var.validate_certificate || local.create_route53_records_only) ? length(local.distinct_domain_names) : 0
 
   zone_id = var.zone_id
   name    = element(local.validation_domains, count.index)["resource_record_name"]
@@ -53,8 +61,7 @@ resource "aws_route53_record" "validation" {
 }
 
 resource "aws_acm_certificate_validation" "this" {
-  provider = aws.acm
-  count    = local.create_certificate && var.validation_method != "NONE" && var.validate_certificate && var.wait_for_validation ? 1 : 0
+  count = local.create_certificate && var.validation_method != "NONE" && var.validate_certificate && var.wait_for_validation ? 1 : 0
 
   certificate_arn = aws_acm_certificate.this[0].arn